#49 - refactor: 캘린더 내역 컴포넌트 분리
